Move test to restful location
